+++ index.php
@@ -11,13 +11,31 @@
     $error = $_GET['error'] ?? '0'; 
 
     // url mode
+
+    if('edit' == $task){
+        if(!hasPrivilege()) {
+            header('location:index.php?task=report');
+            return;
+        }
+    }
+
     if('seed' == $task){
+        if(!isAdmin()) {
+            header('location:index.php?task=report');
+            return;
+        }
         seed();
         $info = "Seeding is Complete...";
     }
 
     // delete url
     if('delete' == $task) {
+        
+        if(!isAdmin()) {
+            header('location:index.php?task=report');
+            return;
+        }
+
         $id = filter_input(INPUT_GET,'id',FILTER_SANITIZE_STRING);
         if($id > 0) {
             deleteStudent($id);
